2016/12/20 Achieved a stable goal version. Program is capable of recognizing CAPS markers and sites where dCAPS markers are possible based on a given Hamming distance threshold.
Plans for the future:
Right now the way it passes information around about possible matches in
the sequences are in lists of indices. I think it might be more editable
long-term and less confusing to read if I modify it to a simple object
that would let me store the Hamming distance along with the index so I
don't have to recompute it later.

Primer melting-temperature calculation is implemented according to the
nearest-neighbor method from  OligoCalc and papers cited by same.

I need to start adding the framework for generating primers.

Also, right now it just straight-up ignores an enzyme if it's present as
an exact match downstream of the assumed primer location. I think it would
be useful to keep information on that enzyme rather than ignoring it. That
way, in the primer generation stage, I can see if it's far enough
downstream that I could generate a diagnostic primer pair before it.

Also, I need to put in an interface for looking at one sequence and trying
to find an enzyme for screening PCR products. I'm not sure if I want the
user to indicate the cut site somehow or if I want to infer the cut site
by looking for a PAM in the sequence. Maybe if I ask them to provide the
specific 20-bp guide?